data update breaks animation
the call of highchartsNGUtils.deepExtend({}, options, {data: null, visible: null});

does not set the options.data to null, therefore the data update was executed through a series.update and not a series.setData -> in the processSeries function. Wrapping an angular.extend around should solve the problem with data and visible attributes.