Use addPoint in series scope watch
Checks if the series scope watch finds only a single data point of a series was altered. If so, uses the Highcharts Series.addPoint function to add the point. Implements logic to see if the point was added as a push and shift operation to take advantage of the shift parameter of the addPoint function.